refactor(frames): merge Gnome/Windows Tabs into FramedTabs
Phase B5 of frame unification. Replace the two near-identical Tabs.svelte
(~165 lines each) with one src/renderer/Panel/frames/FramedTabs.svelte
driven by a `style` prop: List icon/dividers come from getFrameConfig(style),
class prefix is derived (g-/w-), and the per-frame .tabs gap/color/padding
are scoped by [data-frame] so the class stays `.tabs`. Both g-*/w-* :global
blocks are kept verbatim (mutually exclusive at runtime) → pixels identical.

Also harden the signature oracle to ignore Svelte scoped-style hash classes
(svelte-xxxx) — an implementation detail that changes on restructure, not a
visual diff; baselines regenerated. DOM signature byte-identical for both
styles across 3 runs.
